refactoring: presentation
