+++ cmd/app/wire.go
@@ -10,13 +10,13 @@ import (
 
 	"github.com/reangeline/foodscan_backend/internal/domain/contracts/repositories"
 	uc_interface "github.com/reangeline/foodscan_backend/internal/domain/contracts/usecases"
-	"github.com/reangeline/foodscan_backend/internal/validation/protocols"
 
 	"github.com/reangeline/foodscan_backend/internal/domain/usecases"
 	"github.com/reangeline/foodscan_backend/internal/infra/databases"
-	"github.com/reangeline/foodscan_backend/internal/validation/validators"
 
-	"github.com/reangeline/foodscan_backend/internal/infra/http/controllers"
+	"github.com/reangeline/foodscan_backend/internal/presentation/controllers"
+	"github.com/reangeline/foodscan_backend/internal/presentation/validation/protocols"
+	"github.com/reangeline/foodscan_backend/internal/presentation/validation/validators"
 )
 
 var setUserRepositoryDependency = wire.NewSet(
