+++ internal/domain/contracts/usecases/user_usecase_interfaces.go
@@ -9,4 +9,5 @@ import (
 type UserUseCaseInterface interface {
 	CreateUser(ctx context.Context, input *dtos.CreateUserInput) error
 	CheckEmailExists(email string) (bool, error)
+	FindByEmail(email string) (*dtos.UserOutputDTO, error)
 }
