+++ internal/infra/http/routes/user_routes.go
@@ -2,14 +2,14 @@ package routes
 
 import (
 	"github.com/go-chi/chi/v5"
-	"github.com/reangeline/foodscan_backend/internal/infra/http/controllers"
+	"github.com/reangeline/foodscan_backend/internal/presentation/controllers"
 )
 
-func InitializeUserRoutes(uc *controllers.UserController) chi.Router {
+func InitializeUserRoutes(controller *controllers.UserController) chi.Router {
 	r := chi.NewRouter()
 
 	r.Route("/users", func(r chi.Router) {
-		r.Post("/", uc.CreateUser)
+		r.Post("/", controller.CreateUser)
 	})
 
 	return r
