+++ internal/presentation/validation/validators/user_validator.go
@@ -1,10 +1,10 @@
 package validators
 
 import (
-	"errors"
 	"strings"
 
 	"github.com/reangeline/foodscan_backend/internal/dtos"
+	err "github.com/reangeline/foodscan_backend/internal/presentation/errors"
 )
 
 type UserValidator struct{}
@@ -16,19 +16,19 @@ func NewUserValidator() *UserValidator {
 func (uv *UserValidator) ValidateUser(user *dtos.CreateUserInput) error {
 
 	if !strings.Contains(user.Email, "@") {
-		return errors.New("please add a valid email")
+		return err.ErrValidEmail
 	}
 
 	if user.Email == "" {
-		return errors.New("email is required")
+		return err.ErrEmailIsRequired
 	}
 
 	if user.Name == "" {
-		return errors.New("name is required")
+		return err.ErrNameIsRequired
 	}
 
 	if user.LastName == "" {
-		return errors.New("lastname is required")
+		return err.ErrLastNameIsRequired
 	}
 
 	return nil
